# Provide better coverage data in case of implicit exceptions