--- a/org.jacoco.core/src/org/jacoco/core/internal/flow/LabelFlowAnalyzer.java
+++ b/org.jacoco.core/src/org/jacoco/core/internal/flow/LabelFlowAnalyzer.java
@@ -54,6 +54,11 @@ public final class LabelFlowAnalyzer extends MethodVisitor {
 	boolean first = true;
 
 	/**
+	 * Label instance of the last line start.
+	 */
+	Label lineStart = null;
+
+	/**
 	 * Create new instance.
 	 */
 	public LabelFlowAnalyzer() {
@@ -91,6 +96,11 @@ public final class LabelFlowAnalyzer extends MethodVisitor {
 		if (successor) {
 			LabelInfo.setSuccessor(label);
 		}
+	}
+
+	@Override
+	public void visitLineNumber(final int line, final Label start) {
+		lineStart = start;
 	}
 
 	@Override
@@ -174,6 +184,7 @@ public final class LabelFlowAnalyzer extends MethodVisitor {
 			final String name, final String desc, final boolean itf) {
 		successor = true;
 		first = false;
+		markMethodInvocationLine();
 	}
 
 	@Override
@@ -181,6 +192,13 @@ public final class LabelFlowAnalyzer extends MethodVisitor {
 			final Handle bsm, final Object... bsmArgs) {
 		successor = true;
 		first = false;
+		markMethodInvocationLine();
+	}
+
+	private void markMethodInvocationLine() {
+		if (lineStart != null) {
+			LabelInfo.setMethodInvocationLine(lineStart);
+		}
 	}
 
 	@Override