--- a/org.jacoco.doc/docroot/doc/changes.html
+++ b/org.jacoco.doc/docroot/doc/changes.html
@@ -22,6 +22,8 @@
 
 <h3>New Features</h3>
 <ul>
+  <li>Better detection of coverage in code blocks with implicit exceptions.
+      (GitHub <a href="https://github.com/jacoco/jacoco/issues/310">#310</a>).</li>
   <li>Added lifecycle-mapping-metadata.xml for M2E
       (GitHub <a href="https://github.com/jacoco/jacoco/issues/203">#203</a>).</li>
   <li>Allow locales with country and variant for Ant report task
@@ -34,6 +36,12 @@
       <code>destfile</code> attribute is now passed as an absolute path also in
       the default case
       (GitHub <a href="https://github.com/jacoco/jacoco/issues/301">#301</a>).</li>
+</ul>
+
+<h3>API Changes</h3>
+<ul>
+  <li>The exec file version has been updated and is not compatible with previous
+      versions.</li>
 </ul>
 
 <h2>Release 0.7.4 (2015/02/26)</h2>