… jacoco-maven-plugin.test/it/it-report-nomatch/nomatch.exec
… org.jacoco.ant.test/src/org/jacoco/ant/data/nomatch.exec
… org.jacoco.ant.test/src/org/jacoco/ant/data/sample1.exec
… org.jacoco.ant.test/src/org/jacoco/ant/data/sample2.exec