@Godin 2015-05-18
For comparison

this PR:
![pr-310](https://cloud.githubusercontent.com/assets/138671/7692134/f6d29c0e-fdc0-11e4-9184-577f8975b1cc.png)

PR #261:
![pr-261](https://cloud.githubusercontent.com/assets/138671/7692132/f373a2f6-fdc0-11e4-9ea0-4d151d421f8f.png)