@buildhive 2015-05-19
[Java Code Coverage Tools » jacoco #380](https://buildhive.cloudbees.com/job/jacoco/job/jacoco/380/) SUCCESS
This pull request looks good
[(what's this?)](https://www.cloudbees.com/what-is-buildhive)