@marchof 2015-05-19
@Godin Nice, thx! Another useful comparison would be the total number of probes and the execution duration for a reasonable test suite.