@mackerl 2015-05-19
Testsuite with 6000 JUnit Tests, 3 Million LoC under Test, 48% Coverage
without Coverage: 12 hours
with Coverage from Master without PR #261 29 hours
with Coverage from Master with PR #261 30 hours
with Coverage from Master with PR #310 coming soon