@marchof 2015-05-19
@mackerl Wow, 12h for 6000 Unit-Tests? Are this actually *unit* tests? Or do you fork a VM for each test case? This would also explain why the runtime with coverage more than doubles. Jsu ta comparison: JaCoCo runs 1000 Tests with coverage in 80sec:

http://www.eclemma.org/jacoco/trunk/test/index.html

Anyways, many thanks for your ongoing feedback, very helpful!