@mackerl 2015-05-19
@marchof You are welcome. A few hundred of the tests are more on the API Layer. They have a runtime from a few minutes up to one hour. We also do not mock everything - there are actually around 1000 tests involving a real Oracle or SqlServer Database, where a specific datastage is set up during the bootstrap of the Unit Test Framework. Those tests typically run up to a few minutes. Nevertheless the biggest part of our testsuite are real unit tests with a runtime of around a few milli seconds.
Thanks for the build ....  would it be an easy task for you to provide an eclemma update site, where this change in jacoco is included? Otherwise I can also do it by myself