@marchof 2015-05-19
@mackerl What counter does your coverage percentage refer to? Would be nice to know the percentages for onstructions, branches and lines separately.