@marchof 2015-05-19
I will create a comparison for commons-collections tests tonight, camparing the following figures:

* execution time
* covered and total instructions
* covered and total branches
* covered and total lines
* total number of probes
* exec file size