@Godin 2015-05-19
@marchof I'm wondering if possible to update your PR to display lines with exceptions as partially covered as it was done in PR #261 ? IMO this makes report in such cases more readable.