@andrioli 2015-05-19
+1 for @Godin proposal. IMO the partially covered lines make the report more "understandable".