@Godin 2015-05-19
Indeed - I meant "understandable", when wrote "readable" :wink: Thx for correction @andrioli