@Godin 2015-05-19
@marchof out of [InstrumentationBenchmark](https://github.com/Godin/jacoco-benchmarks-playground/blob/master/src/main/java/org/jacoco/benchmarks/InstrumentationBenchmark.java) - comparison of "cold" performance of instrumentation of jar, supposed that offline instrumentation not much different from online ( 10 forks, 1 iteration, no warmup, JDK7u80 ) :

elasticsearch = org.elasticsearch:elasticsearch:1.1.2 , 13M
guava = com.google.guava:guava:18.0 , 2.2M
rt = rt.jar of used JVM , 60M

master:
```
  elasticsearch: 4372.632 ± 142.458 ms/op
          guava: 1763.230 ±  33.403 ms/op
             rt: 8443.019 ± 257.801 ms/op
```

this PR:
```
  elasticsearch: 4373.558 ± 106.255 ms/op
          guava: 1815.460 ±  77.315 ms/op
             rt: 8574.634 ± 435.716 ms/op
```

PR #261:
```
  elasticsearch: 4366.380 ± 167.617 ms/op
          guava: 1822.976 ± 141.603 ms/op
             rt: 8502.746 ± 291.181 ms/op
```

Note that benchmark can be improved to exclude time of jar reading, but IMO it already shows that we can forget about impact of changes on instrumentation time for both PRs.