@Godin 2015-05-19
@marchof and btw I believe that file size computable from number of probes, so why you plan to track it?