@marchof 2015-05-19
Here are the results for commons-collections (14,594 test cases):

Branch | Runtime | Exec File Size | Probes | Instructions      | Branches      | Lines
------ | ------- | -------------- | ------ | ----------------- | ------------- | ---------------
Run    | 5.919s  |                |        |                   |               |
Master | 7.222s  | 58,182         | 15,988 | 109,092 / 130,209 | 6,005 / 8,364 | 25,804 / 30,841
PR 310 | 7.421s  | 59,944         | 29,851 | 109,455 / 130,209 | 6,004 / 8,364 | 25,869 / 30,841
PR 261 | 7.571s  | 62,570         | 50,086 | 111,475 / 130,209 | 6,010 / 8,364 | 26,460 / 30,841

While we see a significant increase in probes between the different approaches, runtime seems not be really effected. Also memory overhead should be minimal as we need 1 byte per probe. As a conclusion probes are pretty cheap and we can actually heavily use them!