@Godin 2015-05-19
@marchof SonarSource C++ Plugin, modular project but all classes included in instrumentation and so in dump files , 48K of source code and 1K of classes not counting tests , 2400 tests and 92.7% coverage as reported by SonarQube with JaCoCo 0.7.4 , single shot, wall clock time, first measurement - "mvn test", subsequent - "mvn jacoco:prepare-agent test" :

without coverage = 1m18.3s
master = 1m36.9s
this PR = 1m36.8s
PR #261 = 1m37.1s

We don't have aggregated JaCoCo reports, so can't quickly provide all other numbers, but can work on this if needed.

Taking into account run-to-run variance I consider that all 3 branches actually demonstrate same time, so changes are insignificant.