@Godin 2015-05-19
@marchof even, if addition of probes has insignificant impact on performance, we anyway should be careful at least due to limits on size of methods :smiling_imp: and what about impact on size of class files on disk and in memory, and impact on size of compiled code in memory? :laughing: