@marchof 2015-05-19
@Godin yes, there are differences, see my table above. The columns Instructions, Branches and Lines show covered vs. total items each.