@marchof 2015-05-19
Regarding coverage status of methods with exceptions: If we spent a probe per method invocation we can even mark those lines as "green".

BTW, there are a couple of other implicit exceptions (NPE, IndexOutOfBounds etc.) which will still not result in coverage.