@Godin 2015-05-19
@marchof my point was that maybe you'll see difference even between multiple runs of exactly the same JaCoCo build