@marchof 2015-05-19
@Godin I just tried with master: The figures seem to be reproducible.