@Godin 2015-05-19
@marchof not sure that I understand your point about "green" lines. Original proposal was to show lines 12, 27, 36, 44, 48 as partially covered (see screenshots). You propose to show them as green aka fully covered? If so, then understandably is the same as in case of red aka not-covered - let's take line 12 as example with slightly changed names of methods:
```
a(); // green
b(); // if green or red, then where exception happened? in method "b" or in method "c"?
     // and in case of yellow it is clear that exception happened in method "b".
c(); // red
```