@Godin 2015-05-19
@marchof ok, so then I suppose that figures are different exactly because of difference of branches on not-covered and partially-covered.